requirements.txt
# AI视频助手 - 后端依赖包
gradio==6.3.0
torch>=2.5.0
torchaudio>=2.5.0
openai-whisper>=20230314

# 视频处理
opencv-python>=4.8.0
moviepy>=1.0.3

# 音频处理
ffmpeg-python>=0.2.0

# 数据处理
numpy==2.2.6
pandas>=2.0.0

# 文本向量和检索
sentence-transformers>=2.2.0
scipy>=1.10.0
transformers>=4.41.0
huggingface-hub>=0.20.0
scikit-learn>=1.7.0
tqdm>=4.60.0

# 系统工具
pathlib2>=2.3.7
argparse

# 日志和配置
pyyaml>=6.0

requests==2.32.5
deep-translator

# Flask和Web服务
Flask>=2.3.0
Flask-CORS>=4.0.0
PyJWT>=2.8.0

# 认证和安全
bcrypt>=4.0.0

# LLM相关依赖
openai>=1.0.0
anthropic>=0.3.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
